infra/cloud_function/requirements.txt
functions-framework==3.*
google-cloud-storage>=2.18
google-cloud-pubsub>=2.23
mlops-agents
requests
